azure_functions_project/requirements.txt
azure-functions>=1.18.0
azure-storage-blob>=12.19.0
pandas>=2.1.0
numpy>=1.24.0
scikit-learn>=1.3.0
imbalanced-learn>=0.11.0
